Struct nbtrs::RegionFile

pub struct RegionFile<T> where T: Read + Seek {
    // some fields omitted
}

A region file

These normally have a .mca extension on disk. They contain up to 1024 chunks, each containing a 32-by-32 column of blocks.

Methods

impl<R> RegionFile<R> where R: Read + Seek

fn new(r: R) -> Result<RegionFile<R>, Error>

Parses a region file

fn get_chunk_timestamp(&self, x: u8, z: u8) -> Option<u32>

Returns a unix timestamp of when a given chunk was last modified. If the chunk does not exist in this Region, return None.

Panics

x and z must be between 0 and 31 (inclusive). If not, panics.

fn chunk_exists(&self, x: u8, z: u8) -> bool

Does the given chunk exist in the Region

Panics

x and z must be between 0 and 31 (inclusive). If not, panics.

fn load_chunk(&mut self, x: u8, z: u8) -> Result<Tag, Error>

Loads a chunk into a parsed NBT Tag structure.

Panics

x and z must be between 0 and 31 (inclusive). If not, panics.